chore: move logic for chart flags to custom hook
